pyproject.toml
[tool.poetry]
name = "instagram-django-server"
version = "0.1.0"
description = ""
authors = ["leedg919 <117734848+leedg919@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "instagram_django_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
